Replaced blocking delay with systick timer
This update includes the systick module, which provides elapsed time measurements, which eliminates the need for blocking delays.  The systick module is not quite right yet because the C run time library is not yet added to the project, so things like run time division operations are not supported.  So, as a step towards that end, the math to compute the systick timer reload value is done by the C preprocessor, and the systick frequency parameter in the set_systick_frequency() function is currently ignored.